refactor(commands): migrate auth (login/logout/relay/auth) to internal/commands
Issue #282 step 5 — thick commands last. Pure relocation: LoginCmd/AuthCmd/
LogoutCmd/RelayCmd, doLogin, oauthDeviceLogin (GitHub device flow with
transient-error retry), syncHostCreds (#172 account-switch semantics), and
openBrowser. restartDaemonFn now points at RestartDaemonAfterLogin from the
lifecycle file; the openBrowserFn/restartDaemonFn test seams are unchanged.

Auth tests move with the code; logoutCmd gains a characterization test for
token+user clearing.
